Get movement list instead of isolated movement
