+++ app/src/main/kotlin/com/adriangl/pokeapi_mvvm/di/Di.kt
@@ -1,6 +1,7 @@
 package com.adriangl.pokeapi_mvvm.di
 
 import android.content.Context
+import com.adriangl.pokeapi_mvvm.moves.MovesApi
 import com.adriangl.pokeapi_mvvm.network.PokeApi
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.adapters.Rfc3339DateJsonAdapter
@@ -49,8 +50,12 @@ val networkModule = Kodein.Module("network") {
 
     bind<PokeApi>() with singleton {
         val retrofit: Retrofit = instance()
-
         retrofit.create(PokeApi::class.java)
+    }
+
+    bind<MovesApi>() with singleton {
+        val retrofit: Retrofit = instance()
+        retrofit.create(MovesApi::class.java)
     }
 }
