+++ app/src/main/kotlin/com/adriangl/pokeapi_mvvm/moves/MovesApi.kt
@@ -1,8 +1,11 @@
 package com.adriangl.pokeapi_mvvm.moves
 
+import com.adriangl.pokeapi_mvvm.network.MoveName
+import com.adriangl.pokeapi_mvvm.network.PokemonMove
 import retrofit2.http.GET
+import retrofit2.http.Path
 
 interface MovesApi {
-    @GET("/api/v2/move/{id}")
-    suspend fun getMove(moveId: MoveId): PokemonMove
+    @GET("/api/v2/move/{name}")
+    suspend fun getMove(@Path("name") moveName: MoveName): PokemonMove
 }
