+++ app/src/main/kotlin/com/adriangl/pokeapi_mvvm/pokemonlist/PokemonListActivity.kt
@@ -2,6 +2,7 @@ package com.adriangl.pokeapi_mvvm.pokemonlist
 
 import android.os.Bundle
 import android.view.View
+import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.SearchView
 import androidx.lifecycle.LiveData
@@ -11,6 +12,7 @@ import com.adriangl.pokeapi_mvvm.R
 import com.adriangl.pokeapi_mvvm.utils.injection.viewModel
 import com.mini.android.toggleViewsVisibility
 import kotlinx.android.synthetic.main.pokemonlist_activity.*
+import mini.onFailure
 import mini.onSuccess
 import org.kodein.di.Kodein
 import org.kodein.di.KodeinAware
@@ -54,6 +56,9 @@ class PokemonListActivity : AppCompatActivity(), KodeinAware {
             resource
                 .onSuccess {
                     pokemonListAdapter.list = it
+                }
+                .onFailure {
+                    Toast.makeText(this, "Error", Toast.LENGTH_SHORT).show()
                 }
         })
     }
